requirements.txt
numpy
pillow
flask
rembg